Bring over openstack-mojo-specs utils
This is an interim step. The openstack-mojo-specs utils need to
be imported into zaza, made python3 compliant and then deprecated in
openstack-mojo-specs.

This PR imports the utils and applies Chris MacNaughton's py3 work. The
original naming has been left in place.

Ultimately, juju_utils will be rendered obsolete as we move over one
function at a time to libjuju. juju_get is the first to get adapted.

I would like a pre-review on the approach, with the understating this
is an incomplete WIP.