Allow quick devlopment testing iterations
Zaza should avoid the failures of the past. Namely, requiring a full
red-deploy for every test run. During development of the tests
themselves this is causes painful turn around times.

This is trivially avoided by adding a couple of extra command line
switches to functest-run-suite.

--no-destroy to avoid destroying what is already
deployed.
--no-deploy to avoid redeploying the existing model.

With these switches we can either upgrade-charm and change settings
with --no-destroy alone or skip model setup entirely
with both --no-destroy and --no-deploy.

The main question at hand is the confusion of the original switch
--keep-model which is intended for use for after the test suite runs.
Do we really need this at all? Should we just default to keeping the
model?

What other naming convention questions are there?