Handle set origin
After the series upgrade a config changed hook will fire with openstack
origin or source still set to ubuntu series n-1 which will Traceback.

It is necessary to pause all primary charms which will then allow the
setting of openstack-origin or source to distro after the first unit is
in progress with the series upgrade.