Juju daemons remain up after pre-series-upgrade
The latest edge version of juju has the juju daemons remain up after
the completion of the pre-series-upgrade hook. This is a slight change
in behavior which requires a change to the zaza process. But it also
allows us to test for model idleness which we could not do when the
agents were down. This should avoid more race conditions.