Allow application_name parameter to config_change
When the charm under test is not the self.application_name of the object
of the setup/test class zaza.test_charms.test_utils.config_change would
try to make a configuration change to the wrong charm. This lead to
https://github.com/openstack-charmers/zaza/issues/168.

This change allows passing application_name to config_change for when
the charm under test is not the same as the object's
self.application_name.

Closes-Issue: https://github.com/openstack-charmers/zaza/issues/168