Use separate tmp dirs for each unit
When running the file ready function each unit would download into the
same tmp directory. For files with restrictive permissions this would
lead to a permission denied and false negative on the second unit. i.e.
file is 444 and the subsequent scp fails due to the inability to
overwrite the first file.

Use a separate tmp dir for each unit to avoid this problem.