Switch masakari tests to use pacemaker-remote (#203)
Switch masakari tests to use pacemaker-remote. For this the method
of simulating a host failure has to change to sending the process a
sigterm as shuting down pacemaker-remote any other way is seen as an
orderly shutdown and does not trigger things like stonith.

Also add check to ensure all nodes appear to be online from a
crm pov before starting failover test.