ceph/rbd_mirror: Optionally require presence of images in wait-helper
When you tell Ceph to resync mirrored RBD images it will in
practice remove and re-create the image.

At present the image state wait helper will happilly accept no
images in a pool as a positive outcome.

Add optional ``require_images_in`` parameter that allows the wait
helper to block even when no images are available in the pool (yet)