Remove double call to sync_wrapper
Calling a function via the sync_wrapper should never result in
another function being called via the sync_wrapper. If this happens
then a "RuntimeError: This event loop is already running" exception
is raised. In the example below no model is set via environment
variables. This leads to the following sequence:

1) Call run_action_on_leader which is a call to
   async_run_action_on_leader using thw sync_wrapper.
2) async_run_action_on_leader calls the async function run_in_model
3) run_in_model calls get_juju_model
4) get_juju_model then calls get_current_model which is a call to
   async_get_current_model using thw sync_wrapper.
5) A second call to run_until_complete is issued against the
  existing loop.