Add helpers to get test bundles, tests and config
This change adds helpers to retrieve the model alias mappings for
bundles, configuration and tests as specified in the tests.yaml.
This includes handling configuration written with and without
explicit model alias maps. See below for background.

To support cross-model-relations test authors need the ability to
express which tests/config or bundles should be targetted at which
model. To do this zaza will support model aliases. This will allow
an author to associate a model alias with a bundle. Subsequent
configuration and test steps can then be targetted at that model
alias. This aslo supports multiple bundles to be deployed in a test
run and each associated with its own model alias.

Backward compatability with existing config which does not specify a
model alias is achieved by mapping objects which do not specify a
model alias to a default alias.

It is possible for an author to use both approaches in the same
tests.yaml eg:

    charm_name: vault
    configure:
    - zaza.openstack.charm_tests.vault.setup.basic_setup
    - vault_model:
      - zaza.openstack.charm_tests.vault.setup.basic_setup
    - keystone_model:
      - zaza.openstack.charm_tests.ks.setup.user_setup
    gate_bundles:
    - xenial-ha-mysql
    - vault_model: xenial-ha-mysql
      keystone_model: keystone-certs
    tests:
    - zaza.openstack.charm_tests.vault.tests.VaultTest
    - vault_model:
      - zaza.openstack.charm_tests.vault.tests.VaultTest
    - keystone_model:
      - zaza.openstack.charm_tests.keystone.tests.ProjTest