Allow bundles to disable the local overlay.
If a bundle does not use the charm referenced in tests.yaml then
no local overlay  should be used. This change allows a bundle to
explicitly disable a local overlay.

Longer term the overlay handling in zaza is expected to be
rewritten and this change is really a tactical fix until that
happens.