[WIP] Add model alias support to configure & test steps
Support for model aliases in the tests.yaml recently landed. This
works when all steps are run together using functest-run-suite.
However, the command line utilities for running each step individually
did not support using model aliases.

This change allows the '-m/-model' argument to be used to pass in
a map in the format '-m model_alias:existing_model'. This
argument can be specified multiple times to pass in multiple
maps.