Use absolute path to charm in default local overlay
Starting with Juju version 2.7, relative charm paths will be
interpreted in relation to the location of the overlay file.

Previous versions would interpret paths relative to the location
of the main bundle file.

Build an absolute path for use in the default local overlay so we
can work with both paradigms.