Add timeout for juju controllers refresh
If you have a dead controller, which is already deleted but still has a
local record in juju, it will cause a few problems without hints:

1. juju destroy-model will freeze forever
2. juju kill-model will fail with unuseful message
3. juju controllers --refresh will freeze forever

The only way to fix so far is `juju unregister`.

A related bug has been reported:

    https://bugs.launchpad.net/juju-core/+bug/1593506

`functest-prepare` will also freeze forever because of 3, without any
message. To avoid wasting time on waiting, add timeout and help message
for people to find the problem earlier.

Signed-off-by: Joe Guo <guoqiao@gmail.com>