Fix pid checks and make them consistent.
The pid checks need to quote the service name to ensure service
with characters such as brackes do not break the tests eg
'aodh-evaluator: AlarmEvaluationService worker(0)'

Also, the pid checks need to be able to handle the situation where a
service has more than one process. Currently the code partially fails
when using pgrep:

```
$ pgrep -f 'apache2'
16120
16126
16130
$ pgrep -f 'apache2' | cut -f1 -d ' '
16120
16126
16130
$ stat -c %Y /proc/$(pgrep -f "apache2" | cut -f1 -d ' ')
1569235446
stat: cannot stat '16126': No such file or directory
stat: cannot stat '16130': No such file or directory
```

When using pidof the first in the list is chosen, which may not be
correct:

```
$ pidof -x 'apache2'
16130 16126 16125 16124 16120
$ pidof -x apache2 | cut -f1 -d ' '
16130
$ stat -c %Y /proc/$(pidof -x apache2 | cut -f1 -d ' ' | cut -f1 -d ' ')
1569235472
```

This change updates both approaches to return the start time of the
oldest process.