Add --force for deploy to override juju constraint checks
To be able to run tests on a series that juju doesn't support, the
--force param is needed for the 'juju deploy' command.  This PR
adds the option to the deploy phase and the test runner.  It is
intentionally not provided in the bundle as the intention is for it to
be obvious (and more difficult) to override Juju constraint checks.