Add model_name args to juju utils functions
The functions in zaza.utilities.juju were missing the model_name
argument. This is the first part of the process of deprecating
zaza.openstack.utilities.juju in favour of the duplicate
functions in zaza.utilities.juju. The corresponding functions in
zaza.openstack.utilities.juju already have the model_name arg.

As part of testing it became apparent that is_subordinate_application
was broken so this PR contains a fix for that.

All the functions in zaza.utilities.juju were tested locally with a
model that was in focus and one that was not: http://paste.ubuntu.com/p/bNmzkXKbsx/