Add maas support to get_unit_name_from_host_name
get_unit_name_from_host_name currently assumes that the machine
number of a host can be derived from the last part of the hostname.
This is not true of maas machines. The machine number can be found
recursing over the machines in the model and looking for a matching
display name. To keep things exciting display_name is an empty
string when using the openstack provider.