Remove domain name when mapping hostname to units
Remove the domain name when trying to derive the machine number
from the hostname. This bug was introduced when the duplicate
`get_unit_name_from_host_name` functions were consolidated and
a wrapper put around the z-o-t version to point at this one. The
domain name was correctly removed by the old z-o-t version.