Strip whitespace from the passed bundle for func-test-runner
This is so that it's a bit more rubust when passing model_alias:bundle
into the func-test-runner as most written combinations in tests.yaml
have a space (as it's a 'hash' in the yaml) but really, it's treated as
a string when passed directly.  We probably should use something other
that ':' as the divider as that means something in yaml.