Change the async_get_status() so that it caches
This change is to reduce the number of async_get_status() calls to juju
to limit it to every 'n' seconds.  Multiple different async co-routines
will access the same returned status within an 'n' second period if the
additional 'refresh' param is passed as False.  Otherwise the co-routine
will stall for the interval and then fetch a new result.  The idea is
that 'normally' tests are actually not concurrent and the get_status
will either fetch immediately, or wait 'n' seconds if previously
requested quickly.

However, for concurrent co-routines (in particular concurrent series
upgrade tests), multiple co-routines can access the same get_status when
upgrading different machines and charms.