Fix return code checking in check_output_logging
If the child process had not exited before the returncode was checked
then check_output_logging would fail to catch the returncode of
the child process. This is because popen.poll() returns None if the
child process is still running *1.  This change adds a wait to ensure
that the child process is complete before checking the returncode.
Closes issue #342

*1 https://docs.python.org/3/library/subprocess.html#subprocess.Popen.poll