Add force and max_wait parameters to destroy_model()
For tests we need to make sure the model is destroyed, and that means
provided force to it.  We also want to only wait about 10 minutes.
Finally, we want to make sure that the model is actually destroyed.  If
not, we'll bail out as something rather drastic has gone wrong.