Cope with the juju delay on removing model
It turns out that the juju destroy_model() function returns before the
model has dissapeared from .list_models().  It can also still be seen in
the juju status after the function has returned.  Async issues, I
suspect.  Anyway, this commit adds a retry to provide time for the
deleted model to disappear, and thus (hopefully) all the resources to be
cleaned up.