Add force and max_wait parameters to destroy_model() (#409)
* Add force and max_wait parameters to destroy_model()

For tests we need to make sure the model is destroyed, and that means
provided force to it.  We also want to only wait about 10 minutes.
Finally, we want to make sure that the model is actually destroyed.  If
not, we'll bail out as something rather drastic has gone wrong.

* Cope with the juju delay on removing model

It turns out that the juju destroy_model() function returns before the
model has dissapeared from .list_models().  It can also still be seen in
the juju status after the function has returned.  Async issues, I
suspect.  Anyway, this commit adds a retry to provide time for the
deleted model to disappear, and thus (hopefully) all the resources to be
cleaned up.