Really destroy instances when destroying model on openstack provider
Essentially, the destroy-model with --force (via libjuju) doesn't reap
the instances leading to CI failures if multiple bundles are used as it
relies on proper clean-up of the instances and their vips.

This patch uses the undercloud OpenStack directly to remove the
instances that used to belong to the model that was just destroyed.
Only has an affect on the openstack provider.