Add support for running a CMR deployment via the --bundle flag
When running a cross-model Zaza test, it is necessary to
be able to specify two (or more) different bundles in the
same command. This change adds support for multiple bundles,
along with their associated model aliases, to be passed in
and run correctly. This brings the --bundle argument inline
with the --dev, --gate, and --smoke deployment methods.