Fix broken destroy model functionality (#430)
This fixes the destroy method and cleans up after pending tasks.

Also remove unused tenacity import

Tenacity was removed as a) it introduced another point for async code to
run (which was really confusing to debug) and b) seemed overly complex
for the retries that are needed.  This simplifies the retry code to make
it very obvious where the async handovers are.