Add support for guessing the model alias
When calling  functest-run-suite --keep-model --bundle $bundle,
Zaza will read the tests.yaml to determine the model alias to
use for a deployment, unless a model alias is explicitly passed
in with a colon. Additionally, if multiple bundles in the
tests.yaml match, Zaza will log that and fall back to the
default alias.