Handle libjuju disconnects by avoiding model.block_until
The libjuju model.block_until() raises
websockets.exceptions.ConnectionClosed() if the model disconnects for
any reason during the call.  This is extremely annoying, but it is what
it is.

What is actually needed, is to trap the condition, and if it happens,
reconnect to the model and then try the conditions again, as they may
rely on the model being connected.

Related-Bug: zaza: #402