Add regex feature to workload-message and refactor
This patch adds a workload-status-message-regex option for matching with
application unit status messages.  This allows for an accurate match to
empty messages ("^$") as well as matching messages that have additional,
dynamic, information (e.g. queue sizes, network ports, etc.).

It also refactors the check_unit_workload_status_message() function so
that it checkes idle of each unit along with the status and
status_message.  This allows the model to reach 'idle' more organically
and doesn't require everything to be static before checking status
messages.  This results in less output and a slightly quicker check (as
it's not constrained by update-status).