Add global options (#442)
At some point the tests.yaml grew a section called "tests_options",
which was introduced to cope with providing additional, optional, values
to test functions.  This only worked if a tests.yaml is provided; there
was no way to provide options on the command line using functest-test
(for example).

This patch does several things:

 * Adds a read-only module for Dictionaries and Lists.
 * Adds a global options module with helpers to get, set and merge
   optios into the a global options set.
 * Loads the options in from the tests.yaml:tests_options key.
 * Provides a way to pass an option using on the command line using the
   functest-test CLI entry point.

Incidentally, this patch was driven by a need to prevent the explosion
of named tests in a bundle to cover multiple versions of OpenStack.

There is also a minor addition to the zaza.model module to get the lead
unit object.