Fix applications with 0 units in wait_for_application_states
So (async_)wait_for_application_states(..) had a hidden issue before it
was refactored to its current form: essentially, a subordinate charm
with no units would be "ready" as the principle's may not have been
installed 'enough' for Juju to allocate the units for the subordinates.

The refactor of the function to do all the checking in parallel (rather
than waiting on the first, then the 2nd, etc.) would see those as ready
immediately and not check them again.

This refactor allows the caller of the function to explicitly allow
charms to be ready with no units, otherwise every application needs at
least 1 unit for it to be ready.