Fix sleep in async_wait_for_application_states
With the async sleep at the end of the while loop in
async_wait_for_application_states(), it's possible for the loop to miss
the sleep due to the use of `continue` statements in the while block.
When this happens, no other async coroutines will make progress which
means that the status of the juju model will not be updated.  This can
lead to the effect of the function not thinking units have achieved idle
status, when in fact they have.  This leads to erroneous abandoning of a
test run.

Moving the sleep to the start of the while loop ensures that the event
loop will make progress and the status of the model will be updated in
python-libjuju.