Update z.u.installers module for system level systemd
The previous version used --user level systemd services; however, these
are inconsistent as they can stop when the user logs out.  Thus, change
it to system-level service files.  The rest of the changes are to
support installing a module for a user other than the ssh user, as long
as that user has sudo privileges.