Add initial zaza events framework (no tests)
This adds the framework for event logging in zaza.  This provides a
facility to have a pluggable set of event log systems as a collection
which enables event logs from multiple sources to be combined into a
single time-series event stream.  The logs are compatible (and designed)
to be uploaded into InfluxDB. (The upload isn't in this commit).