Add logging plugin to provide event logs from tests
This plugin in is to provide a standardised event loggging facility from
tests.  It essentially provides named (and a default) event logger,
which allows prefilling of fields to produce a primarily line protocol
format log (for InfluxDB).

It plugs into the the zaza.events.Collection class so that all the logs
from multiple places can be collated into a single time-series event
stream.