Add notifcation pub-sub module
This module provides a simple pub-sub style notification system for
zaza.

So, in order to decouple a thing happening, and whether it should be
logged as an event, a simple notification system is introduced in
zaza/notifications/.  This allows a subscriber to subscribe to events,
and for notifications to be placed in the zaza test life-cycle code
(e.g. before and after deploy, before configuration, around functions,
etc) so that the broadbased events of the zaza test lifecycle can be
subscribed to.

This is initially going to be used for the TimeSeries events system, but
the report module will be re-written to subscribe to these events.