Add reset() method to events/plugins
The reset() method is used to reset the collection (and all plugins
associated with that collection).  This is to allow the collection to be
re-used for a subsequent test run.

In zaza, multiple models can be run (although this will be discouraged
at some point), and so the collection needs to be re-used after each
test run (EnviromentDeploy).