Set the MAX_FRAME_SIZE to 256MB (#466)
Due to https://github.com/juju/python-libjuju/issues/458 we've put
a few workarounds into zaza to deal with libjuju disconnections.
Setting the MAX_FRAME_SIZE to something big is supposed to help.  Thus,
set it to 256MB which should deal with most use cases.  Note that the
parallel series-upgrade tests in z-o-t set it to 4GB manually to resolve
the error, but that's probably overkill.

* Xenial now requires force deploy

xenial is not supported by the magpie charm anymore.
For this patchset, just force deploy on first

* Ensure code coverage is met with patch

Previously, the model.py:deployed() function had no tests.  Thus
modifying a line in it causes patch coverage to drop.  Add tests to
solve this little dilemna.