Switch the local overlay default to False
When zaza was first implemented, every bundle was local to a charm, and
it was convenient to assume that the local charm under test should be
used automatically.

Since then, the space has changed, charmcraft has introduced built
artifacts (<name>.charm) and automatic overlays get in the way of this.

This patch switches the default for the local overlay to be False.  This
means that bundles that *want* to use a local overlay have to be
explicit and enable it using "local_overlay_enabled: true".

This means that *most* overlays don't need the key (going forwards).