Enable resolve of install hooks for network issues
Allow for failed install hooks to be resolved in order to provide a set
of tolerance for various network blips and dns issues that may occur in
the networks of the substrates.

This introduces environment variable TEST_MAX_RESOLVE_COUNT which
specifies the maximum number of attempts to resolve an install hook
failure. This defaults to 0 and is should be an integer value.

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>