CI: fix juju crashdump collection
The ``--keep-last-model`` literally only keeps the last model, so
if the failure is in the middle model it will be destroyed.  Use
``--keep-faulty-model`` instead.

The `juju status` command is currently run after tox in the same
job. In the event of non-zero return from tox the shell will stop
execution and subsequently not issue the ``juju status``.  Move
to crashdump job instead.

The ``juju crashdump`` command is currently not scoped to the
Zaza model(s) and subsequently fails to collect data.