machine_os: Add helper to check if unit is in VM
We do our testing accross multiple substrates.  For some features
a substrate may be supported only for testing and in that case it
does not make sense for the charm to encode test-only
functionality.  For those cases we want the test code to prepare
the environment to allow the general purpose code to function.

Add helper to detect if unit is running in a VM.