machine_os: Add helper to toggle VFIO IOMMU mode
In production you would normally change the kernel command line
to configure a systems IOMMU to be in pass-through mode.  This
is done so that adapters used for DPDK or SR-IOV do not need to
translate memory addresses when performing DMA transfers.

When running simulated tests ontop of virtual machines this
becomes more involved as to really test with IOMMU the virtual
machine definition would need to be augmented to add virtual
IOMMU devices, and the host system needs to support this etc.

There is a test mode available in the VFIO driver which allow
you to run without using the IOMMU.  This is unsafe to do, but
it is useful for testing.  Since this is something we do not
want people to do in the wild, we keep this as a test-framework
function rather than a feature of any charm itself.