Document requirement for syncing files with z-o-t
The ``zaza`` and ``zaza-openstack-tests`` projects are related,
and currently the latter is installed as a package inside the
former.  As a consequence ``zaza-openstack-tests`` needs to
carry a copy of ``zaza/__init__.py`` as this file will be
overwritten by the copy in ``zaza-openstack-tests`` on install.

Document this fact in the file so that we hopefully remember
this the next time we update it.