Find subordinates reading from top level tree.
This patch changes the way subordinates are filtered, since juju-2.9
doesn't populate the 'charm' key within the 'subordinates' key in the
principal, so instead this change uses the unit id to get the app name
and read the 'charm' value from the 'applications' section.

This functionality is behind a feature a flag, if the environment
variable TEST_ZAZA_BUG_LP1987332 is set to anything that evaluates to
true, then this new code path is executed.

Related-Bug: #1987332