Add back in the run() function to zaza/__init__.py
Prior to the PR that enabled the async functions to all be
run in a background thread, there was a `run()` function that took
a list of `async def` functions that were run in order and then
the result of the last one returned.

This PR adds back that function.  It works slightly differently, but
should provide the same functionality.