Add support for ignore_hard_deploy_errors (#581)
Add support for the ignore_hard_deploy_errors option. If this is
set to true then charms going into an error state during deployment
will not be fatal. This is useful with k8s charms which can
transition into an error state if a container image takes time to
download.