Add functest-run-module command for centralised tests
This command allows an arbitrary module to be loaded and then a function
to be run that is passed any remaining parameters from the command.
This allows changes to be be to both the command and the runner without
having to modify zaza.  It is going to be used by the centralised tests
stream of work.