Add explicit .readthedocs.yaml
Due to [1] this patch adds an explicit .readthedocs.yaml file for the
readthedocs.org sphinx build. It also fixes some more recent linting
failures in the actual docs that sphinx seems to care about more now.

Finally, it experimentally uses pip-compile (manually) to generate a
doc-requirements.txt from a doc-requirements.in just for the docs. This
has been added to the tox.ini and the rest adjusted accordingly.

[1] https://blog.readthedocs.com/doctool-without-configuration-file/