Set pin-priority for proposed pocket to 500
As of lunar, a lower pin priority is set by default for the proposed
pocket [1]. The lower pin-priority prevents installation of proposed
packages, unless they are selectively installed per package [2]. This
is the case even if the proposed package version is greater than any
other available versions.

This change restores the previous behavior. By creating an apt
preferences file that sets the pin-priority of the proposed pocket to
500, all proposed packages will be installed when proposed is enabled,
assuming the proposed package version is greater than any other
available versions.

[1] https://lists.ubuntu.com/archives/ubuntu-devel/2022-November/042338.html
[2] sudo apt-get install packagename/lunar-proposed