Only wait if an object is awaitable.
Between juju 2.x and 3.x, the resulting return objects from the
run method has changed to be awaitable or not. This change checks
to see if the resulting object can be awaited or not. Without this,
zaza will simply hang waiting on the results of an object which is
not actually awaitable.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>