Update status of action results when unknown
Actions are awaited on and then completed but the status may still
be 'unknown' even with a completion timestamp. Update the action
status by querying the model in this case.

This change also includes the removal of the isawaitable check on
the unit.run_action change in the previous commit. It is not necessary
to wait on the unit.run_action as the contract did not change between
Juju 2.x and 3.x.

Additionally, the isawaitable check was added to all locations that
invoke the unit.run(...) method.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>