juju_utils: fix assumptions about 'subordinate-to'
Few places in the code made assumption that for subordinate charms,
the 'subordinate-to' list would contain only principal charms. However
that is not the case. If a suboridnate charm is related to other
subordinate charms, they show up in each other's 'subordinate-to' lists.

This was especially problematic in the 'get_machines_for_application'
function which would end up in an infinite loop.

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>