Add retry mechanism for the Launchpad API
Quite often Launchpad API is temporarily unavailable. In such cases, without this patch,
a request to get all Ubuntu releases from Launchpad fails immediately, causing the
entire test to fail. For example for the charm-octavia it can take up to 2 hours before
the functional tests get to this point. And if it fails because of the Launchapad API
unavailability, the whole tests suite needs to be restarted. In order to save time and
resources, a simple request retry mechanism is introduced by this patch.

Signed-off-by: Marcin Wilk <marcin.wilk@canonical.com>